Fix nasty bug
Happens when connection is dropped while some of the pieces have already been downloaded and when a new connection with the faker is made, it has none of the remaining pieces. This is a scenario that only really affects the testing, because of the way the faker is written in a single-thread mode and doesn't simulate the asynchronous acquisition of new pieces.